+++ CHANGELOG.md
@@ -1,13 +1,18 @@
 # Changelog
 
-
 ## [v2.3.6.dev0]
 
 ### Added
 - [Storage] Added MinIO storage backend
 - [Core] Allow to pass function args as part of the invocation payload in FaaS backends
+- [Core] Optimize call_async() calls with an internal function caching system
+- [AWS Lambda] New invoke method that reduces total invocation time
+- [Core] Allow to set the bucket name at storage backend level in config
+- [localhost] stop running jobs processes on ctrl-c or exception
+- [AWS Lambda] region_name parameter in config
 
 ### Changed
+- [Core] Improved performance and efficiency of the lithops cleaner background process
 - [AWS Lambda] Use layer from Klayers API for pre-compiled Amazon Linux numpy binaries
 
 ### Fixes
@@ -17,6 +22,11 @@
 - [Core] Fixed customized_runtime feature
 - [Core] Fixed get_result() execution after wait() when throw_except is set to False
 - [Core] Fixed internal executions
+- [Core] Fixed 'lithops storage list' CLI when a bucket is empty
+- [Standalone] Fixed execution
+
+### Deleted
+- [Core] Deleted invoke_pool_threads param from map and map_reduce calls. Now it must be set at backend level in config
 
 
 ## [v2.3.5]